requirements.txt
pillow
pdf2image
pyzbar